requirements.txt
flask
flask-login
h5py
pandas
numpy
pillow